+++ src/main/java/blackjack/controller/BlackJackController.java
@@ -7,8 +7,8 @@ import blackjack.domain.CardDeck;
 import blackjack.domain.Dealer;
 import blackjack.domain.Gambler;
 import blackjack.domain.Player;
-import blackjack.domain.PlayerDto;
-import blackjack.domain.PlayersDto;
+import blackjack.dto.PlayerDto;
+import blackjack.dto.PlayersDto;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 import java.util.List;
